Add menu and set a basic cytoscapejs interface up.
